extract useReference hooks from ReferenceFieldController
