+++ packages/ra-core/src/controller/field/index.ts
@@ -1,9 +1,11 @@
 import ReferenceArrayFieldController from './ReferenceArrayFieldController';
 import ReferenceFieldController from './ReferenceFieldController';
 import ReferenceManyFieldController from './ReferenceManyFieldController';
+import useReference from './useReference';
 
 export {
     ReferenceArrayFieldController,
     ReferenceFieldController,
+    useReference,
     ReferenceManyFieldController,
 };
